認証・認可機能を実装

+++ app/UseCases/Article/DeleteArticleUseCase.php
@@ -3,6 +3,7 @@
 namespace App\UseCases\Article;
 
 use App\Domain\Interfaces\ArticleRepositoryInterface;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteArticleUseCase
 {
@@ -10,8 +11,17 @@ class DeleteArticleUseCase
         private ArticleRepositoryInterface $repository
     ) {}
 
-    public function execute(int $id): void
+    public function execute(int $id, int $currentUserId): void
     {
+        $article = $this->repository->findById($id);
+        if (!$article) {
+            throw new \Illuminate\Database\Eloquent\ModelNotFoundException();
+        }
+
+        if ($article->userId !== $currentUserId) {
+            throw new AccessDeniedHttpException('この記事をする削除する権限がありません。');
+        }
+
         $this->repository->delete($id);
     }
 }
